+++ torchdr/distance/__init__.py
@@ -1,0 +1,15 @@
+from .base import pairwise_distances, symmetric_pairwise_distances_indices
+from .torch import pairwise_distances_torch, LIST_METRICS_TORCH
+from .keops import pairwise_distances_keops, LIST_METRICS_KEOPS
+from .faiss import pairwise_distances_faiss, LIST_METRICS_FAISS
+
+__all__ = [
+    "pairwise_distances",
+    "symmetric_pairwise_distances_indices",
+    "pairwise_distances_torch",
+    "pairwise_distances_keops",
+    "pairwise_distances_faiss",
+    "LIST_METRICS_TORCH",
+    "LIST_METRICS_KEOPS",
+    "LIST_METRICS_FAISS",
+]
